requirements.txt
requests>=2.28.0
certifi>=2024.0.0
gspread>=5.12.0
google-auth>=2.22.0
